requirements.txt
altgraph==0.17.3
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
cryptography==3.4.6
diffq==0.2.3
Dora==0.0.3
einops==0.6.0
future==0.18.2
julius==0.2.7
kthread==0.2.3
librosa==0.9.2
llvmlite==0.39.1
natsort==8.2.0
numba==0.56.4
numpy==1.23.4
omegaconf==2.2.3
opencv-python==4.6.0.66
onnx
onnxruntime-gpu==1.16.1
Pillow==9.3.0
pip==22.2.2
psutil==5.9.4
pydub==0.25.1
pyglet==1.5.23
pyperclip==1.8.2
pyrubberband==0.3.0
pytorch_lightning==2.0.0
PyYAML==6.0
resampy==0.2.2
scipy==1.9.3
soundfile==0.11.0
soundstretch==1.2
torch==1.13.1
tqdm
urllib3==1.26.12
wget==3.2
samplerate==0.1.0
screeninfo==0.8.1
PySoundFile==0.9.0.post1; sys_platform != 'windows'
SoundFile==0.9.0; sys_platform == 'windows'
